Restructure to enable addition of further commands
