+++ cmd/init/templates/.gitignore
