+++ cmd/init/templates/README.md
