+++ cmd/init/templates/repos.txt
