+++ cmd/root.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	initCmd "github.com/skyscanner/turbolift/cmd/init"
 	"github.com/spf13/cobra"
 	"log"
 )
@@ -9,6 +10,10 @@ var rootCmd = &cobra.Command{
 	Use:   "turbolift",
 	Short: "Turbolift",
 	Long:  `Mass refactoring tool for repositories in GitHub`,
+}
+
+func init() {
+	rootCmd.AddCommand(initCmd.CreateInitCmd())
 }
 
 func Execute() {
