+++ go.mod
@@ -3,7 +3,6 @@ module github.com/skyscanner/turbolift
 go 1.16
 
 require (
-	github.com/briandowns/spinner v1.12.0
 	github.com/fatih/color v1.7.0
 	github.com/spf13/cobra v1.1.3
 	github.com/stretchr/testify v1.7.0
